Refactoring. Resolving auth issue
